Refactor some location queries
