+++ dtbase/core/queries.py
@@ -16,6 +16,7 @@ from dtbase.core.structure import (
     LocationSchemaIdentifierRelation,
     LocationStringValue,
 )
+from dtbase.core import utils
 
 
 def location_identifiers_by_schema(session):
@@ -39,3 +40,25 @@ def location_identifiers_by_schema(session):
         )
     )
     return query
+
+
+def select_location_by_coordinates(schema_name, session, **kwargs):
+    schema_sq = location_identifiers_by_schema(session).subquery()
+    schema_q = session.query(
+        schema_sq.c.identifier_id,
+        schema_sq.c.identifier_name,
+        schema_sq.c.identifier_datatype,
+    ).where(schema_sq.c.schema_name == schema_name)
+    identifiers = session.execute(schema_q).fetchall()
+    location_q = session.query(Location.id)
+    for id_id, id_name, id_datatype in identifiers:
+        value_class = aliased(utils.get_value_class_from_type_name(id_datatype))
+        location_q = location_q.join(
+            value_class,
+            and_(
+                value_class.location_id == Location.id,
+                value_class.value == kwargs[id_name],
+                value_class.identifier_id == id_id,
+            ),
+        )
+    return location_q
